reducing entities to meet free limits
Trying to migrate this back to the SIL managed, free Auth0 tenant
`silid`.  To do so, we have to limit the test entities to 10 (clients,
rules, etc). We don't need them all as they are there are placeholders
for testing and reflection.  I'm going to try this out, which may
require a few tries...